requirements.txt
geopandas==0.12.0
numpy==1.23.4
pandas==1.5.0
python-decouple==3.6
SQLAlchemy==1.4.41
